Increase key length storage capacity to u64
Uses `u64` instead of `u32` when serializing key lengths in test
utilities. This allows support for larger keys and prevents potential
overflow when dealing with long sequences.

+++ llm-runner/src/transformer/model.rs
@@ -1223,7 +1223,7 @@ mod tests {
 
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1429,7 +1429,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1495,7 +1495,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1551,7 +1551,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1673,7 +1673,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1808,7 +1808,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1885,7 +1885,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -1930,7 +1930,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
